Add files via upload
The call-center dataset basically has over 32,900 records of data that describes calls made to various call centers. It includes the ID of the call, duration of the call-in minutes, the name of the person who called, their satisfaction score and many other attributes that you will see as we go.